BitCoinStreamData/requirements.txt
google-cloud-firestore
google-cloud-storage
google-cloud-pubsub
websockets
python-dotenv
apache-beam[gcp]
google-cloud-bigquery
firebase-admin